+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### Fixed
 - CI build always builds docker images from scratch
   [#823](https://github.com/nextcloud/cookbook/pull/823) @christianlupus
+- Fix test script after update in docker-compose
+  [#833](https://github.com/nextcloud/cookbook/pull/833) @christianlupus
 
 
 ## 0.9.6 - 2021-10-18